backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-socketio>=5.11.0
python-chess>=1.9.4
redis>=5.0.3
python-telegram-bot>=21.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
greenlet>=3.0.0
